pyproject.toml
[tool.poetry]
name = "mtl-scrapmeup"
version = "0.1.0"
description = ""
authors = ["Mariusz Smenżyk <mariusz.smenzyk@bravelab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.45.1"
spotipy = "^2.25.1"
fpdf2 = "^2.8.3"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
